requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.28.1
google-genai>=0.8.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
modal==1.3.4
pinecone>=3.0.0
voyageai>=0.2.0
gitpython>=3.1.40
aiofiles>=23.0.0
